main_pc_code/requirements.txt
# Core dependencies
pyzmq>=25.1.0
numpy>=1.24.0
torch>=2.0.0
scikit-learn>=1.3.0
psutil>=5.9.0

# Logging and monitoring
python-json-logger>=2.0.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
mypy>=1.0.0
flake8>=6.0.0

# Core Dependencies
flask==2.3.3
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.4.2
fastapi==0.103.1
uvicorn==0.23.2
websockets==12.0
aiohttp==3.9.1
asyncio==3.4.3
trio==0.23.1
zmq==0.0.0
pyzmq==27.0.0

# Audio Processing (Alternative packages to avoid compilation issues)
librosa==0.10.1
sounddevice==0.4.6
soundfile==0.12.1
# pyaudio==0.2.13  # Commented out - requires compilation
noisereduce==2.0.1
webrtcvad==2.0.10
pydub==0.25.1
ffmpeg-python==0.2.0
openai-whisper==20231117
pvporcupine==3.0.0

# Machine Learning & AI
torchaudio==2.1.0
transformers==4.34.0
sentencepiece==0.1.99
accelerate==0.23.0
bitsandbytes==0.41.1
optimum==1.13.2
peft==0.5.0
trl==0.7.4
einops==0.7.0
xformers==0.0.22.post7

# Face Recognition (Alternative packages)
insightface==0.7.3
opencv-python-headless==4.8.1.78
opencv-python==4.8.1.78
mediapipe==0.10.7
# dlib==19.24.2  # Commented out - requires compilation

# System & Monitoring
GPUtil==1.4.0
nvidia-ml-py3==7.352.0
prometheus-client==0.17.1
mss==9.0.1

# Database
sqlalchemy==2.0.21
alembic==1.12.0
asyncpg==0.28.0
redis==5.0.1
psycopg2-binary

# Vision Processing
Pillow==10.0.1
scikit-image==0.21.0
scipy==1.11.3

# Utilities
tqdm==4.66.1
colorama==0.4.6
python-magic==0.4.27
watchdog==3.0.0
googletrans==4.0.0-rc1

# Testing
pytest-asyncio==0.21.1
